refactor(import): split the Disqus importer into a core and an adapter
`src/lib/disqus-import.ts` was one 505-line file doing three separate
jobs: converting Disqus' comment HTML to markdown, parsing Disqus' XML
export, and writing normalised threads/comments/users into D1. Only the
middle one is Disqus-specific. Issue #104 wants adapters for Remark42,
Comentario, isso and Cusdis, and every one of them needs the third job
verbatim.

Split along those seams, no behaviour change:

  src/lib/import/core.ts             the source-agnostic writer, plus the
                                     normalised SourceThread/SourceComment/
                                     SourceAuthor shape and the
                                     ImportAdapter contract
  src/lib/import/disqus.ts           parseDisqusXml + DISQUS_ADAPTER
  src/lib/import/html-to-markdown.ts htmlToMarkdown, shared

htmlToMarkdown lands in the core rather than the Disqus adapter because
it is not Disqus-shaped: Remark42's export carries rendered HTML in
`text` and leaves the markdown `orig` field empty for every comment it
imported itself, so that adapter needs the same conversion. It is not a
sanitizer — the core's one sanitising step is still renderMarkdown over
the strict allowlist in src/lib/markdown.ts.

Two SQL statements move from interpolated literals to binds so the
source name is the adapter's, not a hard-coded "disqus". The values are
identical for the Disqus path, and the bind positions the existing
suite asserts on are unchanged. `runDisqusImport` stays as a thin
wrapper over `runImport(db, DISQUS_ADAPTER, …)`, so the admin route,
the operator page and scripts/import-disqus.ts keep their call shape.

Refs #105
